# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 紅白予想 API (kohaku-yoso)
  version: 1.0.0
  description: >
    NHK紅白歌合戦の出場者・歌唱曲・結果を予想するWebアプリのREST API。
    実装は Cloudflare Workers + Hono。詳細は docs/API設計書.md を参照。
    認証はセッションCookie(JWT/HS256)。ブラウザからは credentials:include で送る。
servers:
  - url: /
    description: 同一オリジン（本番は SPA と同居）
...
paths:
...
  /api/admin/seasons:
    post:
      tags: [admin]
      summary: シーズン作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [year]
              properties:
                year: { type: integer, minimum: 1950, maximum: 2100 }
                title: { type: string }
                predictionOpenAt: { type: string, format: date-time }
      responses:
        "201":
          description: 作成
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Season" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/seasons/{id}/close:
    post:
      tags: [admin]
      summary: 締切操作（公式発表日時の設定）
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [announcedAt]
              properties:
                announcedAt:
                  type: string
                  format: date-time
                  description: 公式発表の日時。過去日時も可。
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Season" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/results:
    post:
      tags: [admin]
      summary: 結果確定（冪等）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [seasonId, entries]
              properties:
                seasonId: { type: string }
                entries:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [artistId, appeared]
                    properties:
                      artistId: { type: string }
                      appeared: { type: boolean }
                      songId: { type: string, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  count: { type: integer }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/artists:
    post:
      tags: [admin]
      summary: アーティスト作成（別名同時可）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                nameKana: { type: string }
                genderGroup: { type: string }
                officialUrl: { type: string, format: uri }
                aliases:
                  type: array
                  items: { type: string }
      responses:
        "201":
          description: 作成
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Artist" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/songs:
    post:
      tags: [admin]
      summary: 曲作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [artistId, title]
              properties:
                artistId: { type: string }
                title: { type: string }
                titleKana: { type: string }
                releaseYear: { type: integer, minimum: 1900, maximum: 2100 }
      responses:
        "201":
          description: 作成
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Song" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/external/artists:
    get:
      tags: [admin]
      summary: 外部アーティスト検索（登録補助）
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ExternalArtist" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/external/tracks:
    get:
      tags: [admin]
      summary: 外部曲検索（登録補助）
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ExternalTrack" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
...
    Season:
      type: object
      properties:
        id: { type: string }
        year: { type: integer }
        title: { type: string, nullable: true }
        predictionOpenAt: { type: string, nullable: true }
        predictionCloseAt: { type: string, nullable: true, description: "NULL=受付中" }
        resultConfirmedAt: { type: string, nullable: true }
        isOpen: { type: boolean, description: "受付中フラグ" }
    Artist:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        nameKana: { type: string, nullable: true }
        aliases:
          type: array
          items: { type: string }
    Song:
      type: object
      properties:
        id: { type: string }
        artistId: { type: string }
        title: { type: string }
        titleKana: { type: string, nullable: true }
...
    ExternalArtist:
      type: object
      properties:
        source: { type: string, enum: [spotify, musicbrainz] }
        externalId: { type: string }
        name: { type: string }
        imageUrl: { type: string, nullable: true }
        url: { type: string, nullable: true }
        detail: { type: string, nullable: true }
    ExternalTrack:
      type: object
      properties:
        source: { type: string, enum: [spotify, musicbrainz] }
        externalId: { type: string }
        title: { type: string }
        artistName: { type: string }
        releaseYear: { type: integer, nullable: true }
...
  responses:
    ValidationError:
      description: 入力が不正（VALIDATION_ERROR）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Forbidden:
      description: 権限不足（FORBIDDEN）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: リソースが存在しない（NOT_FOUND）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
security:
  - sessionCookie: []